📌 chore: mongo manager (DoS)

+++ database/manager.js
@@ -14,18 +14,27 @@ module.exports = class Manager {
 	}
 
 	static async findServer(id) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		const result = await serverModel.findOne({ serverId: id });
 
 		return result;
 	}
 
 	static async getPrefix(id) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		const result = await serverModel.findOne({ serverId: id });
 
 		return result.prefix;
 	}
 
 	static async updateServerPrefix(id, prefix) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		const result = await serverModel.findOne({ serverId: id });
 
 		if (result) {
@@ -37,6 +46,9 @@ module.exports = class Manager {
 	}
 
 	static async createUser(id) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		const user = new userModel({
 			userId: id,
 		});
@@ -46,12 +58,18 @@ module.exports = class Manager {
 	}
 
 	static async getUserById(id) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		let user = await userModel.findOne({ userId: id });
 		if (!user) { user = await this.createUser(id); }
 		return { user };
 	}
 
 	static async updateUserById(id, data) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		if (typeof data !== 'object') { throw Error('\'data\' must be an object'); }
 		const user = await this.getUserById(id);
 		if (!user) { await this.addUser(id); }
@@ -60,10 +78,16 @@ module.exports = class Manager {
 	}
 
 	static async removeUser(id) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		await userModel.findOneAndDelete({ userId: id });
 	}
 
 	static async updateServerById(id, settings) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		if (typeof settings !== 'object') { throw Error('\'settings\' must be an object'); }
 
 		const guild = await this.findServer(id);
@@ -74,6 +98,9 @@ module.exports = class Manager {
 	}
 
 	static async removeServer(id) {
+		if (typeof id !== 'string' || id.length !== 24) {
+			throw new Error('Invalid ID');
+		}
 		await serverModel.findOneAndDelete({ guildId: id });
 	}
 };
